requirements.txt
astroid==2.3.3
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.2
bs4==0.0.1
cffi==1.13.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.2
matplotlib==3.1.2
mccabe==0.6.1
mip==1.6.6
network==0.1
networkx==2.4
numpy==1.18.0
ortools==7.4.7247
pandas==0.25.3
parsel==1.5.2
Protego==0.1.16
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
queuelib==1.5.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
Twisted==19.10.0
typed-ast==1.4.1
virtualenv==16.7.9
w3lib==1.21.0
wrapt==1.11.2
zope.interface==4.7.1
